OO-291: fix wrong mapping of the business group view, all counter to long

+++ src/main/java/org/olat/group/model/BusinessGroupViewImpl.java
@@ -79,7 +79,7 @@ public class BusinessGroupViewImpl extends PersistentObject implements BusinessG
 		return numOfParticipants;
 	}
 
-	public void setNumOfParticipants(int numOfParticipants) {
+	public void setNumOfParticipants(long numOfParticipants) {
 		this.numOfParticipants = numOfParticipants;
 	}
